Move playlist item out to it's own model and intiialize it's table
